+++ R/gisco_get_healthcare.R
@@ -1,4 +1,4 @@
-#' Get locations of healthcare services in Europe.
+#' Get locations of healthcare services in Europe
 #'
 #' @concept infrastructure
 #' @family infrastructure