+++ tests/testthat/test-utils.R
@@ -44,10 +44,10 @@ test_that("Errors on database", {
 
 
   expect_message(
-    n <- gsc_api_cache("https://www.dhh.this.is.fake/",
+    n <- gsc_api_cache("https://github.com/dieghernan/a_fake_thing_here",
       verbose = FALSE
     ),
-    "https://www.dhh.this.is.fake/  not reachable."
+    "404"
   )
 
   expect_null(n)